% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cML_estimate_DP.R
\name{cML_estimate_DP}
\alias{cML_estimate_DP}
\title{Estimate with With Data Perturbation}
\usage{
cML_estimate_DP(
  b_exp,
  b_out,
  se_exp,
  se_out,
  K,
  num_pert = 200,
  random_start = 0,
  maxit = 100
)
}
\arguments{
\item{b_exp}{Vector of estimated effects for exposure.}

\item{b_out}{Vector or estimated effects for outcome.}

\item{se_exp}{Vector of standard errors for exposure.}

\item{se_out}{Vector of standard errors for outcome.}

\item{K}{Constraint parameter, number of invalid IVs.}

\item{num_pert}{Number of perturbation, default is 200.}

\item{random_start}{Number of random starting points, default is 0.}

\item{maxit}{Maximum number of iteration.}
}
\value{
A list contains:
theta_v is the vector estimated thetas,
se_v is vector of standard errors of estimated thetas,
l_v is vector of negative log-likelihood. Vectors all have length
num_pert.
}
\description{
With multiple perturbed data, get estimated theta,
se of estimated theta
and negative log-likelihood,
using multiple random starting points.
}
