% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cML_estimate_pt.R
\name{cML_estimate_pt}
\alias{cML_estimate_pt}
\title{Estimate with Regular Likelihood Based on Multiple Permutations}
\usage{
cML_estimate_pt(
  b_exp,
  b_out,
  se_exp,
  se_out,
  K,
  num_permutation = 100,
  initial_theta = 0,
  initial_mu = rep(0, length(b_exp)),
  maxit = 100
)
}
\arguments{
\item{b_exp}{Vector of estimated effects for exposure.}

\item{b_out}{Vector or estimated effects for outcome.}

\item{se_exp}{Vector of standard errors for exposure.}

\item{se_out}{Vector of standard errors for outcome.}

\item{K}{Constraint parameter, number of invalid IVs.}

\item{num_permutation}{Number of permutations.}

\item{initial_theta}{Starting point for theta.}

\item{initial_mu}{Starting point for mu.}

\item{maxit}{Maximum number of iteration.}
}
\value{
A list contains: theta, estimated causal effect;
se_theta, standard error of theta;
Neg_l, minus log likelihood.
}
\description{
Estimate theta, standard error of theta, negative likelihood
with constrained maximum likelihood, averaged from multiple permutations.
}
