% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_cML.R
\name{mr_cML}
\alias{mr_cML}
\title{Title}
\usage{
mr_cML(
  b_exp,
  b_out,
  se_exp,
  se_out,
  n,
  K_vec = 0:(length(b_exp) - 2),
  random_start = 0
)
}
\arguments{
\item{random_start}{}
}
\value{

}
\description{
Title
}
